Renamed Favorites and NewMeetup for consistency in naming for pages + cleaner imports for pages
