+++ src/pages/NewMeetupPage.js
@@ -1,4 +1,4 @@
-import NewMeetupForm from "../components/meetups/NewMeetupForm";
+import { NewMeetupForm } from '../components';
 
 export default function NewMeetupsPage() {
   return (
