+++ src/pages/index.js
@@ -1,0 +1,3 @@
+export { default as AllMeetupsPage } from './AllMeetupsPage';
+export { default as FavoritesPage } from './FavoritesPage';
+export { default as NewMeetupPage } from './NewMeetupPage';
